+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+* Change state to gray when unrestricting ([issue](https://github.com/M66B/XPrivacy/issues/1190))
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.20 BETA**